# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API de Tarefas",
        "description": "API para gerenciar tarefas, incluindo criação, atualização e conclusão de tarefas.",
        "contact": {
            "name": "Felipe Vieira Andrade",
            "email": "felipebcv@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/token": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Gerar Token de Acesso",
                "description": "Realiza o login do usuário e retorna um token de acesso.",
                "operationId": "37ef70106dbe77944f68df793e7a9de9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "teste@px.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "@3das$dasst341#08787jh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "description": "Retrieves all tasks for the authenticated user.",
                "operationId": "89f841aadff57d91e6136a738fd4a0f6",
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "description": "Creates a new task for the authenticated user.",
                "operationId": "5325d35634686a28458a46c33dbf99dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task by ID",
                "description": "Fetches a specific task by its ID.",
                "operationId": "9c708295c919ce91e53ecdcc78244841",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "description": "Updates the details of a specific task.",
                "operationId": "10a24bf5d2e53be95928af7d28fd0e70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/{id}/start": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Start a task",
                "description": "Starts the task, setting its status to 'In Progress'.",
                "operationId": "ca9851001dd17e75efda7c2f2d83b709",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task started successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/{id}/complete": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Complete a task",
                "description": "Marks the task as 'Completed'.",
                "operationId": "641817d0f886b8d2939551bcb7dc3c4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task completed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Task",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Título da Tarefa Exemplo"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descrição da tarefa exemplo"
                    },
                    "status": {
                        "type": "string",
                        "example": "Pendente"
                    },
                    "priority": {
                        "type": "string",
                        "example": "Média"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    }
                },
                "type": "object"
            },
            "TaskStoreRequest": {
                "title": "Requisição para Criar Tarefa",
                "required": [
                    "title",
                    "description",
                    "status",
                    "priority"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Título da Nova Tarefa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descrição detalhada da tarefa"
                    },
                    "status": {
                        "type": "string",
                        "example": "Pendente"
                    },
                    "priority": {
                        "type": "string",
                        "example": "Média"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    }
                },
                "type": "object"
            },
            "TaskUpdateRequest": {
                "title": "Requisição para Atualizar Tarefa",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Título Atualizado da Tarefa"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descrição atualizada"
                    },
                    "status": {
                        "type": "string",
                        "example": "Em Andamento"
                    },
                    "priority": {
                        "type": "string",
                        "example": "Alta"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        },
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}